moved ConfigurationFactory & ConfigurationFactory to phpDocumentor namespace
