+++ src/phpDocumentor/ConfigurationFactory.php
@@ -1,8 +1,6 @@
 <?php
 
-namespace phpDocumentor\Configuration;
-
-use phpDocumentor\Uri;
+namespace phpDocumentor;
 
 final class ConfigurationFactory
 {
@@ -170,7 +168,7 @@ final class ConfigurationFactory
         $domElement = $dom->importNode($domElement, true);
         $dom->appendChild($domElement);
 
-        $dom->schemaValidate(__DIR__ . '/phpdoc.xsd');
+        $dom->schemaValidate(__DIR__ . '/configuration/phpdoc.xsd');
 
         $error = libxml_get_last_error();
